renamed redis to node_rights
